No-Country-Back-End/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.5.0
djangorestframework==3.15.2
gunicorn==23.0.0
itsdangerous==2.2.0
packaging==24.1
pillow==10.4.0
psycopg==3.2.3
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.7.0
